[IMPROVE] 코드 리팩터링
[공통]
 - SpringBoot 3.2.2 로 update 및 기타 라이브러리 최신화

[UpbitClient]
 - 응답값 중 Double 을 모두 BigDecimal 로 교체 (정밀 계산)

[UniversalUtilAndDao]
 - 공통접근 DB (Oracle, InfluxDB) 공통모듈로 분리
 - 주문매도상태 SellType 공통화
 - Util 성 함수 공통모듈로 분리

[Manager]
 - 패키지 depth 변경 (dev.lutergs.santa.trade -> dev.lutergs.santa.trade.manager)
 - UpbitClient 의 BigDecimal 대응
 - Kafka Topic 을 Spring 환경변수에서 가져오도록 변경

[Worker]
 - UpbitClient 의 BigDecimal 대응

+++ TradeWorker/src/main/kotlin/dev/lutergs/santa/trade/worker/domain/entity/Alarm.kt
@@ -1,6 +1,7 @@
 package dev.lutergs.santa.trade.worker.domain.entity
 
 import dev.lutergs.upbitclient.api.exchange.order.OrderResponse
+import java.math.BigDecimal
 import java.util.*
 
 
@@ -31,7 +32,7 @@ data class TradeResultMessage(
 
 data class TradeResultMsgKey(
   val uuid: UUID,
-  val profit: Double
+  val profit: BigDecimal
 )
 
 data class TradeResultMsgValue(
