+++ TradeWorker/src/main/kotlin/dev/lutergs/santa/trade/worker/domain/entity/TradeResult.kt
@@ -1,7 +1,9 @@
 package dev.lutergs.santa.trade.worker.domain.entity
 
+import dev.lutergs.santa.universal.oracle.SellType
 import dev.lutergs.upbitclient.api.exchange.order.OrderResponse
 import dev.lutergs.upbitclient.dto.OrderType
+import java.math.BigDecimal
 
 class TradeResult {
   val buy: OrderResponse
@@ -41,7 +43,7 @@ class TradeResult {
     }
   }
 
-  private fun earnPrice(): Double {
+  private fun earnPrice(): BigDecimal {
     // 매도 주문이 존재하는지 검사
     this.sell ?: throw IllegalStateException("매도 주문이 없습니다. 가격을 계산할 수 없습니다. 전체 값 : $this")
     return this.sell.totalPrice() - this.buy.totalPrice() - (this.sell.paidFee + this.buy.paidFee)
