+++ UpbitClient/src/main/kotlin/dev/lutergs/upbitclient/api/exchange/order/c_Orders.kt
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize
 import dev.lutergs.upbitclient.dto.NumberStringDeserializer
 import dev.lutergs.upbitclient.dto.OffsetDateTimeDeserializer
 import dev.lutergs.upbitclient.dto.UuidDeserializer
+import java.math.BigDecimal
 import java.time.OffsetDateTime
 import java.util.UUID
 
@@ -38,24 +39,24 @@ data class OrdersResponse(
   @JsonProperty("side") val side: String,
   @JsonProperty("ord_type") val ordType: String,
   @JsonDeserialize(using = NumberStringDeserializer::class)
-  @JsonProperty("price") val price: Double,
+  @JsonProperty("price") val price: BigDecimal,
   @JsonProperty("state") val state: String,
   @JsonProperty("market") val market: String,
   @JsonDeserialize(using = OffsetDateTimeDeserializer::class)
   @JsonProperty("created_at") val createdAt: OffsetDateTime,
   @JsonDeserialize(using = NumberStringDeserializer::class)
-  @JsonProperty("volume") val volume: Double,
+  @JsonProperty("volume") val volume: BigDecimal,
   @JsonDeserialize(using = NumberStringDeserializer::class)
-  @JsonProperty("remaining_volume") val remainingVolume: Double,
+  @JsonProperty("remaining_volume") val remainingVolume: BigDecimal,
   @JsonDeserialize(using = NumberStringDeserializer::class)
-  @JsonProperty("reserved_fee") val reservedFee: Double,
+  @JsonProperty("reserved_fee") val reservedFee: BigDecimal,
   @JsonDeserialize(using = NumberStringDeserializer::class)
-  @JsonProperty("remaining_fee") val remainingFee: Double,
+  @JsonProperty("remaining_fee") val remainingFee: BigDecimal,
   @JsonDeserialize(using = NumberStringDeserializer::class)
-  @JsonProperty("paid_fee") val paidFee: Double,
+  @JsonProperty("paid_fee") val paidFee: BigDecimal,
   @JsonDeserialize(using = NumberStringDeserializer::class)
-  @JsonProperty("locked") val locked: Double,
+  @JsonProperty("locked") val locked: BigDecimal,
   @JsonDeserialize(using = NumberStringDeserializer::class)
-  @JsonProperty("executed_volume") val executedVolume: Double,
+  @JsonProperty("executed_volume") val executedVolume: BigDecimal,
   @JsonProperty("trades_count") val tradesCount: Int
 )
