+++ UpbitClient/src/main/kotlin/dev/lutergs/upbitclient/api/quotation/ticker/TickerRequester.kt
@@ -6,6 +6,7 @@ import dev.lutergs.upbitclient.api.RequestDao
 import dev.lutergs.upbitclient.webclient.Requester
 import dev.lutergs.upbitclient.dto.*
 import reactor.core.publisher.Flux
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalTime
 
@@ -28,25 +29,25 @@ data class TickerResponse(
   @JsonDeserialize(using = TimeDeserializer::class)
   @JsonProperty("trade_time_kst") val tradeTimeKst: LocalTime,
   @JsonProperty("trade_timestamp") val tradeTimestamp: Long,
-  @JsonProperty("opening_price") val openingPrice: Double,
-  @JsonProperty("high_price") val highPrice: Double,
-  @JsonProperty("low_price") val lowPrice: Double,
-  @JsonProperty("trade_price") val tradePrice: Double,
-  @JsonProperty("prev_closing_price") val previousClosingPrice: Double,
+  @JsonProperty("opening_price") val openingPrice: BigDecimal,
+  @JsonProperty("high_price") val highPrice: BigDecimal,
+  @JsonProperty("low_price") val lowPrice: BigDecimal,
+  @JsonProperty("trade_price") val tradePrice: BigDecimal,
+  @JsonProperty("prev_closing_price") val previousClosingPrice: BigDecimal,
   @JsonProperty("change") val change: Change,
-  @JsonProperty("change_price") val changePrice: Double,
-  @JsonProperty("change_rate") val changeRate: Double,
-  @JsonProperty("signed_change_price") val signedChangePrice: Double,
-  @JsonProperty("signed_change_rate") val signedChangeRate: Double,
-  @JsonProperty("trade_volume") val tradeVolume: Double,
-  @JsonProperty("acc_trade_price") val accTradePrice: Double,
-  @JsonProperty("acc_trade_price_24h") val accTradePrice24h: Double,
-  @JsonProperty("acc_trade_volume") val accTradeVolume: Double,
-  @JsonProperty("acc_trade_volume_24h") val accTradeVolume24h: Double,
-  @JsonProperty("highest_52_week_price") val highest52weekPrice: Double,
+  @JsonProperty("change_price") val changePrice: BigDecimal,
+  @JsonProperty("change_rate") val changeRate: BigDecimal,
+  @JsonProperty("signed_change_price") val signedChangePrice: BigDecimal,
+  @JsonProperty("signed_change_rate") val signedChangeRate: BigDecimal,
+  @JsonProperty("trade_volume") val tradeVolume: BigDecimal,
+  @JsonProperty("acc_trade_price") val accTradePrice: BigDecimal,
+  @JsonProperty("acc_trade_price_24h") val accTradePrice24h: BigDecimal,
+  @JsonProperty("acc_trade_volume") val accTradeVolume: BigDecimal,
+  @JsonProperty("acc_trade_volume_24h") val accTradeVolume24h: BigDecimal,
+  @JsonProperty("highest_52_week_price") val highest52weekPrice: BigDecimal,
   @JsonDeserialize(using = DateWithHyphenDeserializer::class)
   @JsonProperty("highest_52_week_date") val highest52weekDate: LocalDate,
-  @JsonProperty("lowest_52_week_price") val lowest52weekPrice: Double,
+  @JsonProperty("lowest_52_week_price") val lowest52weekPrice: BigDecimal,
   @JsonDeserialize(using = DateWithHyphenDeserializer::class)
   @JsonProperty("lowest_52_week_date") val lowest52weekDate: LocalDate,
   @JsonProperty("timestamp") val timestamp: Long
